#259 implemented issue and refactor code
