+++ src/main/java/com/a2/a2_automation_system/sportsmanpayments/SportsmanPaymentRepository.java
@@ -1,4 +1,4 @@
-package com.a2.a2_automation_system.tariff;
+package com.a2.a2_automation_system.sportsmanpayments;
 
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
