+++ src/test/java/com/a2/a2_automation_system/user/UserServiceTest.java
@@ -4,19 +4,15 @@ import com.a2.a2_automation_system.group.Group;
 import com.a2.a2_automation_system.group.GroupRepository;
 import com.a2.a2_automation_system.parent.ParentRepository;
 import com.a2.a2_automation_system.relationship.RelationshipRepository;
-import com.a2.a2_automation_system.tariff.SportsmanPayment;
-import com.a2.a2_automation_system.tariff.SportsmanPaymentRepository;
+import com.a2.a2_automation_system.sportsmanpayments.SportsmanPayment;
+import com.a2.a2_automation_system.sportsmanpayments.SportsmanPaymentRepository;
 import com.a2.a2_automation_system.userparam.UserParam;
-import com.a2.a2_automation_system.userparam.UserParamDTO;
 import com.a2.a2_automation_system.userparam.UserParamRepository;
 import org.assertj.core.api.Assertions;
-import org.junit.Assert;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.stubbing.OngoingStubbing;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
